refactor: 테스트 코드에 final 키워드 추가
final 키워드를 추가함으로써 안정성 보장

+++ src/main/java/racingcar/validator/TotalRoundsValidator.java
@@ -23,7 +23,7 @@ public class TotalRoundsValidator {
 
     private static void validateIsNumber (String rawTotalRounds){
         try {
-            Float.parseFloat(rawTotalRounds);
+            Double.parseDouble(rawTotalRounds);
         } catch (NumberFormatException e) {
             throw new IllegalArgumentException(NOT_NUMBER.getMessage());
         }
